refactor(site): adopt Page/PageView pattern and shared header
- Rename SettingsPageContent → AgentSettingsPageView (git mv)
- Split AnalyticsPageContent into AgentAnalyticsPage (data fetching)
  and AgentAnalyticsPageView (pure render), delete the old file
- Extract the shared expand-sidebar/logo/chime header bar into
  AgentPageHeader, used by create, settings, and analytics pages
- Simplify AgentCreatePage by replacing inline header JSX with
  AgentPageHeader
